+++ druid-spring-boot-starter/src/main/java/com/alibaba/druid/spring/boot/autoconfigure/DruidStatProperties.java
@@ -1,0 +1,176 @@
+/*
+ * Copyright 1999-2017 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.druid.spring.boot.autoconfigure;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * @author lihengming<89921218@qq.com>
+ */
+@ConfigurationProperties("spring.datasource.druid")
+public class DruidStatProperties {
+    private String[] aopPatterns;
+    private StatViewServlet statViewServlet = new StatViewServlet();
+    private StatFilter statFilter = new StatFilter();
+
+
+    public String[] getAopPatterns() {
+        return aopPatterns;
+    }
+
+    public void setAopPatterns(String[] aopPatterns) {
+        this.aopPatterns = aopPatterns;
+    }
+
+    public StatViewServlet getStatViewServlet() {
+        return statViewServlet;
+    }
+
+    public void setStatViewServlet(StatViewServlet statViewServlet) {
+        this.statViewServlet = statViewServlet;
+    }
+
+    public StatFilter getStatFilter() {
+        return statFilter;
+    }
+
+    public void setStatFilter(StatFilter statFilter) {
+        this.statFilter = statFilter;
+    }
+
+    public static class StatViewServlet {
+        private String urlPattern;
+        private String allow;
+        private String deny;
+        private String loginUsername;
+        private String loginPassword;
+        private String resetEnable;
+
+        public String getUrlPattern() {
+            return urlPattern;
+        }
+
+        public void setUrlPattern(String urlPattern) {
+            this.urlPattern = urlPattern;
+        }
+
+        public String getAllow() {
+            return allow;
+        }
+
+        public void setAllow(String allow) {
+            this.allow = allow;
+        }
+
+        public String getDeny() {
+            return deny;
+        }
+
+        public void setDeny(String deny) {
+            this.deny = deny;
+        }
+
+        public String getLoginUsername() {
+            return loginUsername;
+        }
+
+        public void setLoginUsername(String loginUsername) {
+            this.loginUsername = loginUsername;
+        }
+
+        public String getLoginPassword() {
+            return loginPassword;
+        }
+
+        public void setLoginPassword(String loginPassword) {
+            this.loginPassword = loginPassword;
+        }
+
+        public String getResetEnable() {
+            return resetEnable;
+        }
+
+        public void setResetEnable(String resetEnable) {
+            this.resetEnable = resetEnable;
+        }
+    }
+
+    public static class StatFilter {
+        private String urlPattern;
+        private String exclusions;
+        private String sessionStatMaxCount;
+        private String sessionStatEnable;
+        private String principalSessionName;
+        private String principalCookieName;
+        private String profileEnable;
+
+        public String getUrlPattern() {
+            return urlPattern;
+        }
+
+        public void setUrlPattern(String urlPattern) {
+            this.urlPattern = urlPattern;
+        }
+
+        public String getExclusions() {
+            return exclusions;
+        }
+
+        public void setExclusions(String exclusions) {
+            this.exclusions = exclusions;
+        }
+
+        public String getSessionStatMaxCount() {
+            return sessionStatMaxCount;
+        }
+
+        public void setSessionStatMaxCount(String sessionStatMaxCount) {
+            this.sessionStatMaxCount = sessionStatMaxCount;
+        }
+
+        public String getSessionStatEnable() {
+            return sessionStatEnable;
+        }
+
+        public void setSessionStatEnable(String sessionStatEnable) {
+            this.sessionStatEnable = sessionStatEnable;
+        }
+
+        public String getPrincipalSessionName() {
+            return principalSessionName;
+        }
+
+        public void setPrincipalSessionName(String principalSessionName) {
+            this.principalSessionName = principalSessionName;
+        }
+
+        public String getPrincipalCookieName() {
+            return principalCookieName;
+        }
+
+        public void setPrincipalCookieName(String principalCookieName) {
+            this.principalCookieName = principalCookieName;
+        }
+
+        public String getProfileEnable() {
+            return profileEnable;
+        }
+
+        public void setProfileEnable(String profileEnable) {
+            this.profileEnable = profileEnable;
+        }
+    }
+}
