--- druid-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,10 +1,0 @@
-{
-  "groups": [
-    {
-      "name": "spring.datasource.druid",
-      "type": "com.alibaba.druid.pool.DruidDataSource",
-      "sourceType": "org.springframework.boot.autoconfigure.jdbc.DruidAutoConfiguration",
-      "sourceMethod": "dataSource(org.springframework.boot.autoconfigure.jdbc.DataSourceProperties)"
-    }
-  ]
-}
