--- druid-spring-boot-starter/src/main/resources/META-INF/spring.provides
@@ -1,1 +1,0 @@
-provides: druid-spring-boot-starter
