--- druid-spring-boot-starter/src/test/java/com/alibaba/boot/druid/demo/DruidDemoApplication.java
@@ -1,16 +1,0 @@
-package com.alibaba.boot.druid.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * druid demo application
- *
- * @author leijuan
- */
-@SpringBootApplication
-public class DruidDemoApplication {
-    public static void main(String[] args) throws InterruptedException {
-        SpringApplication.run(DruidDemoApplication.class, args);
-    }
-}
