+++ druid-spring-boot-starter/src/test/java/com/alibaba/druid/spring/boot/demo/service/UserService.java
@@ -1,0 +1,8 @@
+package com.alibaba.druid.spring.boot.demo.service;
+
+
+import com.alibaba.druid.spring.boot.demo.model.User;
+
+public interface UserService {
+    User findById(Long id);
+}
