+++ druid-spring-boot-starter/src/test/resources/config-template.properties
@@ -1,0 +1,42 @@
+######Druid数据源配置######
+# JDBC配置
+spring.datasource.url= #别名为spring.datasource.druid.url
+spring.datasource.username= #别名为spring.datasource.druid.username
+spring.datasource.password= #别名为spring.datasource.druid.password
+spring.datasource.driver-class-name= #别名为spring.datasource.druid.driver-class-name，非必填项可通过url解析
+# 连接池配置,下面配置说明请参考Druid Github Wiki，配置_DruidDataSource参考配置
+spring.datasource.druid.initialSize=
+spring.datasource.druid.minIdle=
+spring.datasource.druid.maxActive=
+spring.datasource.druid.maxWait=
+spring.datasource.druid.timeBetweenEvictionRunsMillis=
+spring.datasource.druid.minEvictableIdleTimeMillis=
+spring.datasource.druid.validationQuery=
+spring.datasource.druid.testWhileIdle=
+spring.datasource.druid.testOnBorrow=
+spring.datasource.druid.testOnReturn=
+spring.datasource.druid.poolPreparedStatements=
+spring.datasource.druid.maxPoolPreparedStatementPerConnectionSize=
+spring.datasource.druid.filters= #默认值stat，配置多个英文逗号分隔
+
+######Druid监控配置######
+#下面配置说明请参考Druid Github Wiki，配置_配置WebStatFilter
+spring.datasource.druid.WebStatFilter.enabled= #是否启用StatFilter默认值true
+spring.datasource.druid.WebStatFilter.urlPattern=
+spring.datasource.druid.WebStatFilter.exclusions=
+spring.datasource.druid.WebStatFilter.sessionStatMaxCount=
+spring.datasource.druid.WebStatFilter.sessionStatEnable=
+spring.datasource.druid.WebStatFilter.principalSessionName=
+spring.datasource.druid.WebStatFilter.principalCookieName=
+spring.datasource.druid.WebStatFilter.profileEnable=
+#下面配置说明请参考Druid Github Wiki，配置_StatViewServlet配置
+spring.datasource.druid.StatViewServlet.enabled= #是否启用StatViewServlet默认值true
+spring.datasource.druid.StatViewServlet.urlPattern=
+spring.datasource.druid.StatViewServlet.resetEnable=
+spring.datasource.druid.StatViewServlet.loginUsername=
+spring.datasource.druid.StatViewServlet.loginPassword=
+spring.datasource.druid.StatViewServlet.allow=
+spring.datasource.druid.StatViewServlet.deny=
+#下面配置说明请参考Druid Github Wiki，配置_配置WebStatFilter
+spring.datasource.druid.aop-patterns= #Spring监控AOP切入点，如x.y.z.service.*,配置多个英文逗号分隔
+#如果spring.datasource.druid.aop-patterns要代理的类没有定义interface请设置spring.aop.proxy-target-class=true
