--- src/test/java/com/alibaba/druid/spring/boot/autoconfigure/service/UserService.java
@@ -1,8 +1,0 @@
-package com.alibaba.druid.spring.boot.autoconfigure.service;
-
-
-import com.alibaba.druid.spring.boot.autoconfigure.model.User;
-
-public interface UserService {
-    User findById(Long id);
-}
