--- src/test/java/com/alibaba/druid/spring/boot/autoconfigure/web/UserController.java
@@ -1,19 +1,0 @@
-package com.alibaba.druid.spring.boot.autoconfigure.web;
-
-import com.alibaba.druid.spring.boot.autoconfigure.model.User;
-import com.alibaba.druid.spring.boot.autoconfigure.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController  {
-    @Autowired
-    private UserService userService;
-
-    @RequestMapping("/user/{id}")
-    public User list(@PathVariable Long id) {
-        return userService.findById(id);
-    }
-}
