--- src/test/resources/config/application.properties
@@ -1,19 +1,0 @@
-# 数据源基础配置
-spring.datasource.druid.url=jdbc:mysql://localhost:3306/your-db
-spring.datasource.druid.username=your-username
-spring.datasource.druid.password=your-password
-spring.datasource.druid.driver-class-name=com.mysql.jdbc.Driver
-#启用Druid的内置监控页面
-spring.datasource.druid.StatViewServlet.enabled=true
-#启用Web关联监控配置
-spring.datasource.druid.WebStatFilter.enabled=true
-#启用Spring关联监控配置
-spring.datasource.druid.aop-patterns=com.alibaba.druid.spring.boot.autoconfigure.service.*
-#spring.datasource.druid.maxActive=20
-#spring.datasource.druid.filters=stat,wall,log4j
-#spring.datasource.druid.StatViewServlet.loginUsername=admin
-#spring.datasource.druid.StatViewServlet.loginPassword=123456
-# JPA
-spring.jpa.show-sql= true
-spring.jpa.hibernate.ddl-auto=create-drop
-
